+++ build_runner/lib/src/entrypoint/daemon.dart
@@ -70,6 +70,11 @@ class DaemonCommand extends BuildRunnerCommand {
         await server.stop();
         await logSub.cancel();
       });
+      // Clients can disconnect from the daemon mid build.
+      // As a result we try to relinquish resources which can
+      // cause the build to hang. To ensure there are no ghost processes
+      // fast exit.
+      exit(0);
     }
     return 0;
   }