+++ build_runner/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## 1.2.6
 
+<<<<<<< HEAD
 - No longer assumeTty when logging through the daemon command.
 - Update `build_daemon` to version `0.4.0`. 
+=======
+- Ensure the daemon command always exits.
+>>>>>>> master
 
 ## 1.2.5